refactor: move chain state to wallet plugin
