+++ packages/maskbook/src/extension/background-script/EthereumService.ts
@@ -3,4 +3,3 @@ export * from './EthereumServices/network'
 export * from './EthereumServices/provider'
 export * from './EthereumServices/tokenList'
 export * from './EthereumServices/nonce'
-export * from './EthereumServices/chainState'
