+++ packages/maskbook/src/plugins/Wallet/services/index.ts
@@ -6,3 +6,4 @@ export * from './erc1155'
 export * from './assets'
 export * from './transactions'
 export * from './recentTransactions'
+export * from './chain'
